fix: get the label using the label's path
When using axios and the public URL, the requests would be rejected
because we still need some sort of authentication. We have now
refactored that part a bit so that we store the path of the label in the
configured bucket, and we then use the storage API from 'firebase-admin'
to obtain the data.

NOTE: when using the emulators, validation of the download should be
disabled.